+++ database/migrations/2025_12_15_112449_create_orders_table.php
@@ -16,8 +16,8 @@ return new class extends Migration
             $table->foreignId('user_id')
                 ->constrained('users')
                 ->onDelete('cascade');
-            $table->decimal('price', 18);
-            $table->decimal('amount', 18);
+            $table->float('price', 18);
+            $table->float('amount', 18);
             $table->string('symbol', 6);
             $table->tinyInteger('side');
             $table->tinyInteger('status');
